klo-gba-sidebar: move SelectVision to KloGbaSidebar
