+++ brush/src/components/MapEmptyState/index.js
@@ -6,7 +6,7 @@ const MapEmptyState = () => (
   <span className={style.container}>
     <KlonoaWSit height={250} />
     <h2><strong>No vision selected!</strong></h2>
-    <h3>{'Please, select one at the upper left corner and let\'s paint an amazing map ~~'}</h3>
+    <h3>{'Please, select one at the left bar and let\'s paint an amazing map ~~'}</h3>
   </span>
 )
